+++ internal/cmd/company/iam/add.go
@@ -18,6 +18,7 @@ package iam
 import (
 	"github.com/mia-platform/miactl/internal/clioptions"
 	"github.com/mia-platform/miactl/internal/cmd/company/iam/serviceaccount"
+	"github.com/mia-platform/miactl/internal/cmd/company/iam/user"
 	"github.com/spf13/cobra"
 )
 
@@ -31,6 +32,7 @@ add them to the company selected via the flag or context`,
 
 	cmd.AddCommand(
 		serviceaccount.CreateServiceAccountCmd(options),
+		user.AddCmd(options),
 	)
 
 	return cmd