+++ internal/resources/requests.go
@@ -31,6 +31,11 @@ type ServiceAccountRequest struct {
 	PublicKey PublicKey          `json:"publicKey,omitempty"`
 }
 
+type AddUserRequest struct {
+	Email string             `json:"email"`
+	Role  ServiceAccountRole `json:"role"`
+}
+
 type PublicKey struct {
 	Type      string `json:"kty"` //nolint: tagliatelle
 	Use       string `json:"use"` //nolint: tagliatelle